requirements.txt
tkinter
ttkbootstrap==1.12.0
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
pyarrow==14.0.1
openpyxl==3.1.2
python-dotenv==1.0.0
cairosvg==2.7.1
svglib==1.5.1 